:truck: Move tests to app package
