+++ modules/cli/src/test/kotlin/datamaintain/cli/app/TestAppender.kt
@@ -1,4 +1,4 @@
-package datamaintain.cli
+package datamaintain.cli.app
 
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.core.AppenderBase
